save progress

+++ src/services/notification-control/alchemy.class.ts
@@ -9,9 +9,10 @@ export class AlchemyHandler {
   private static webhookId = notifyConfig.getWebhookId();
 
   public static async addAddressesForTracking(traders: ITraderInfo[]): Promise<void> {
+    if (traders.length === 0) return;
     try {
       // Add each trader to the database
-      const traderPromises = traders.map(async (traderInfo) => {
+      for (const traderInfo of traders) {
         const trader = new TraderModel({
           name: traderInfo.name,
           address: traderInfo.address,
@@ -19,26 +20,28 @@ export class AlchemyHandler {
         });
 
         await trader.save(); // Save the trader to the database
-      });
-
-      await Promise.all(traderPromises);
+      }
+      console.log("saved to db.. saving to webhook");
 
       // Add the addresses to webhook, to start getting tracked
       const addresses = traders.map((trader) => trader.address);
       await this.alchemyNotifySDK.notify.updateWebhook(this.webhookId, {
         addAddresses: addresses,
       });
+      console.log("saved to webhook", this.webhookId);
     } catch (err: any) {
       console.error("Error from addAddressesForTracking:\n", err);
     }
   }
 
   public static async removeAddressesForTracking(traders: ITraderInfo[]): Promise<void> {
+    if (traders.length === 0) return;
     try {
       // remove each trader from the database
       for (const trader of traders) {
         await TraderModel.deleteOne({ name: trader.name });
       }
+      console.log("Successfully removed from db.. removing from webhook");
 
       // remove the addresses from webhook, to stop getting tracked
       const addresses = traders.map((trader) => trader.address);
